server/backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cors==1.0.1
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.2
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.16.1
filelock==3.19.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.4
gunicorn==25.1.0
idna==3.10
jwt==1.4.0
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pycparser==2.22
PySocks==1.7.1
requests==2.32.5
requests-file==2.1.0
setuptools==80.9.0
sqlparse==0.5.3
tldextract==5.3.0
urllib3==2.5.0
whitenoise==6.12.0
zope.event==5.1.1
zope.interface==7.2
